backend/requirements.txt
Django>=5.2.1
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
djangorestframework-simplejwt>=5.5.0
channels>=4.0.0
fhir.resources>=8.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
daphne>=4.0.0
